Apply updates from Zast.Project

+++ Mikibot.Crawler/Http/Bilibili/Model/LiveServer/LiveInitInfo.cs
@@ -10,9 +10,9 @@ namespace Mikibot.Crawler.Http.Bilibili.Model.LiveServer
     public struct LiveInitInfo
     {
         [JsonPropertyName("room_id")]
-        public int RoomId { get; set; }
+        public long RoomId { get; set; }
 
         [JsonPropertyName("uid")]
-        public int BId { get; set; }
+        public long BId { get; set; }
     }
 }
